+++ cg/meta/workflow/mip.py
@@ -10,7 +10,7 @@ from cg.constants import COLLABORATORS, COMBOS, MASTER_LIST, Pipeline
 from cg.constants.constants import FileFormat
 from cg.constants.tags import HkMipAnalysisTag
 from cg.exc import CgError
-from cg.io.base import WriteFile, ReadFile
+from cg.io.controller import WriteFile, ReadFile
 from cg.meta.workflow.analysis import AnalysisAPI
 from cg.meta.workflow.fastq import MipFastqHandler
 from cg.models.cg_config import CGConfig
