+++ tests/apps/scout/test_get_causative_variants.py
@@ -5,7 +5,7 @@ import logging
 from cg.apps.scout.scout_export import Variant
 from cg.apps.scout.scoutapi import ScoutAPI
 from cg.constants.constants import FileFormat
-from cg.io.base import ReadStream
+from cg.io.controller import ReadStream
 
 
 def test_get_causative_variants_no_variants(scout_api: ScoutAPI, case_id: str):
