+++ tests/io/test_io_controller.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from cg.constants.constants import FileFormat
-from cg.io.base import ReadFile, WriteFile, ReadStream, WriteStream
+from cg.io.controller import ReadFile, WriteFile, ReadStream, WriteStream
 from cg.models.mip.mip_sample_info import MipBaseSampleInfo
