+++ tests/mocks/balsamic_analysis_mock.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Tuple
 
 from cg.constants.constants import FileFormat
-from cg.io.base import ReadFile
+from cg.io.controller import ReadFile
 from cg.meta.workflow.balsamic import BalsamicAnalysisAPI
