+++ tests/mocks/mip_analysis_mock.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from cg.constants.constants import FileFormat
-from cg.io.base import ReadFile
+from cg.io.controller import ReadFile
 from cg.models.mip.mip_analysis import MipAnalysis
 from cg.models.mip.mip_metrics_deliverables import MIPMetricsDeliverables
